+++ src/appleseed/renderer/modeling/bssrdf/bssrdfsample.h
@@ -45,7 +45,7 @@ class BSSRDFSample
     BSSRDFSample();
 
     const foundation::Basis3d*  m_shading_basis;
-    double                      m_eta;
+    float                       m_eta;
     size_t                      m_channel;
     foundation::Vector2d        m_point;
     double                      m_rmax2;
